@@ -2,9 +2,9 @@ import React from "react";
 import Link from "next/link";
 import {connect} from "react-redux";
 
-export default connect(state => state)(({pathname}) => (
-    <div className="other">
-        <div>Using Next.js default prop in a wrapped component: {pathname}</div>
+export default connect(state => state)(({toe}) => (
+    <div>
+        <div>This is an error page, it also has access to store: {toe}</div>
         <nav>
             <Link href="/"><a>Navigate to index</a></Link>
         </nav>
