@@ -1,13 +1,12 @@
 import React from "react";
 import Link from "next/link";
-import wrapper from "../src";
-import {makeStore} from "./components/store";
+import {connect} from "react-redux";
 
 class Page extends React.Component {
 
     static getInitialProps({store, isServer, pathname, query}) {
 
-        console.log(Page.name, '- 2. Cmp.getInitialProps uses the store to dispatch things, pathname', pathname, 'query', query);
+        console.log('2. Page.getInitialProps uses the store to dispatch things, pathname', pathname, 'query', query);
 
         // If it's a server, then all async actions must be done before return or return a promise
         if (isServer) {
@@ -24,6 +23,7 @@ class Page extends React.Component {
         // If it's a client, then it does not matter because client can be progressively rendered
         store.dispatch({type: 'TICK', payload: 'client'});
 
+        // Some custom thing for this particular page
         return {custom: 'custom client'};
 
     }
@@ -31,10 +31,12 @@ class Page extends React.Component {
     render() {
         // console.log('5. Page.render');
         return (
-            <div>
-                <div>Redux tick: {this.props.tick} (this page)</div>
+            <div className="index">
+                <div>Redux tick: {this.props.tick}</div>
+                <div>Redux toe: {this.props.toe}</div>
                 <div>Custom: {this.props.custom}</div>
                 <Link href="/other"><a>Navigate</a></Link>
+                <Link href="/error"><a>Navigate to error</a></Link>
             </div>
 
         )
@@ -42,8 +44,4 @@ class Page extends React.Component {
 
 }
 
-wrapper.setDebug(true);
-
-Page = wrapper(makeStore, state => state)(Page);
-
-export default Page;
+export default connect(state => state)(Page);
